python/requirements.txt
python-telegram-bot[webhooks,rate-limiter]==21.6
librosa==0.10.1
matplotlib==3.9.2
numpy==1.26.4
soundfile==0.13.1
fastapi==0.115.2
uvicorn==0.31.1
httpx==0.28.1
setuptools>=70.0.0
